Add BAT tests for mixer command-line testing
These tests should be reporting on the structure and correctness of mixer
and its user interface.

This patch also introduces a docker container to do the build, so we do not
have to worry about running this in different environments such as Travis
CI. The tests are thus confined to a strict, well-known workspace,
and we can avoid many dependency problems.

Signed-off-by: Tudor Marcu <tudor.marcu@intel.com>
Signed-off-by: Matthew Johnson <matthew.johnson@intel.com>