Check access permissions on dirs before mounting
This patch adds an access check on each directory in the config file
before mounting it to the container. This prevents the user from
accidentally (or surreptitiously) granting the container access to
directories they shouldn't.

The config parsing for directory extraction code is also refactored to
provide better results and more accurate error messages.

Signed-off-by: Kevin C. Wells <kevin.c.wells@intel.com>