Add unit tests, and run Travis not as root
This patch adds unit tests for directory access checks, directory extraction,
and mount path list reduction.

For these tests to not fail inside the Docker container on Travis, they
need to not be run as root. This patch thus also creates a non-root user
(but with wheelnopw membership), and runs as this user instead.

Signed-off-by: Kevin C. Wells <kevin.c.wells@intel.com>